+++ test/integration_tests/StdForwardListLoggingTest.cpp
@@ -92,6 +92,14 @@ TEST_CASE("std_forward_list_logging")
 
     std::forward_list<int> empt;
     LOG_INFO(logger, "empt {}", empt);
+
+    // Test rvalue references with forward_list
+    std::forward_list<int> rvalue_forward_list = {100, 200, 300};
+    LOG_INFO(logger, "rvalue_forward_list {}", std::move(rvalue_forward_list));
+
+    // Test with temporary forward_list
+    LOG_INFO(logger, "temp_forward_list {}",
+             std::forward_list<std::string>{"temp1", "temp2", "temp3"});
   }
 
   logger->flush_log();
@@ -156,6 +164,12 @@ TEST_CASE("std_forward_list_logging")
 
   REQUIRE(quill::testing::file_contains(
     file_contents, std::string{"LOG_INFO      " + logger_name + "       empt []"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"LOG_INFO      " + logger_name + "       rvalue_forward_list [100, 200, 300]"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"LOG_INFO      " + logger_name + "       temp_forward_list [\"temp1\", \"temp2\", \"temp3\"]"}));
 
   testing::remove_file(filename);
 }