+++ test/integration_tests/StdMultiSetLoggingTest.cpp
@@ -132,6 +132,13 @@ TEST_CASE("std_multiset_logging")
 
     std::multiset<int> empt;
     LOG_INFO(logger, "empt {}", empt);
+
+    // Test rvalue references with multiset
+    std::multiset<int> rvalue_multiset = {100, 200, 300};
+    LOG_INFO(logger, "rvalue_multiset {}", std::move(rvalue_multiset));
+
+    // Test with temporary multiset
+    LOG_INFO(logger, "temp_multiset {}", std::multiset<std::string>{"temp1", "temp2", "temp3"});
   }
 
   logger->flush_log();
@@ -227,6 +234,12 @@ TEST_CASE("std_multiset_logging")
 
   REQUIRE(quill::testing::file_contains(
     file_contents, std::string{"LOG_INFO      " + logger_name + "       empt {}"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"LOG_INFO      " + logger_name + "       rvalue_multiset {100, 200, 300}"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"LOG_INFO      " + logger_name + "       temp_multiset {\"temp1\", \"temp2\", \"temp3\"}"}));
 
   testing::remove_file(filename);
 }